Rename xfaster_tools to the more specific spec_tools.
Now contains only functions for generating and manipulating model power spectra.

Also move expand_qb function to parse_tools since that's where all the matrix
handling code lives.
