+++ xfaster/__init__.py
@@ -6,6 +6,6 @@ from .xfaster_exec import *
 from .xfaster_class import *
 
 from .parse_tools import load_and_parse
-
+from .spec_tools import get_camb_cl
 
 __version__ = "0.1"
