+++ xfaster/spec_tools.py
@@ -6,7 +6,6 @@ import numpy as np
 __all__ = [
     "wigner3j",
     "get_camb_cl",
-    "expand_qb",
     "scale_dust",
 ]
 
@@ -218,32 +217,3 @@ def get_camb_cl(r, lmax, nt=None, spec="total", lfac=True):
     totCL = powers[spec][: lmax + 1, :4].T
 
     return totCL
-
-
-def expand_qb(qb, bin_def, lmax=None):
-    """
-    Expand a qb-type array to an ell-by-ell spectrum using bin_def.
-
-    Arguments
-    ---------
-    qb : array_like, (nbins,)
-        Array of bandpower deviations
-    bin_def : array_like, (nbins, 2)
-        Array of bin edges for each bin
-    lmax : int, optional
-        If supplied, limit the output spectrum to this value.
-        Otherwise the output spectrum extends to include the last bin.
-
-    Returns
-    -------
-    cl : array_like, (lmax + 1,)
-        Array of expanded bandpowers
-    """
-    lmax = lmax if lmax is not None else bin_def.max() - 1
-
-    cl = np.zeros(lmax + 1)
-
-    for idx, (left, right) in enumerate(bin_def):
-        cl[left:right] = qb[idx]
-
-    return cl
